feat: Add item list component and data source
The code changes include adding a new file `ItemList.tsx` under the `client/components/ui` directory. This file contains a React functional component `ItemList` that renders a list of items with the ability to add items to a shopping cart. The component uses data from the `data.json` file located in the `database` directory.